Add blue alliance functionality
